gdp-poc/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
databricks-sql-connector
